pyproject.toml
[tool.poetry]
name = "nonebot-plugin-datastore"
version = "0.6.0a0"
description = "适用于 Nonebot2 的数据存储插件"
authors = ["hemengyang <hmy0119@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/he0119/nonebot-plugin-datastore"
repository = "https://github.com/he0119/nonebot-plugin-datastore"
documentation = "https://github.com/he0119/nonebot-plugin-datastore#readme"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = { extras = ["httpx"], version = "^2.0.0-rc.1" }
nonebot-plugin-localstore = ">=0.2.0,!=0.3.0,!=0.4.0,<0.5.0"
sqlalchemy = "^2.0.3"
aiosqlite = ">=0.17,<0.19"
alembic = "^1.9.1"

anyio = { version = ">=3.6", optional = true }
click = { version = ">=8.0", optional = true }
typing-extensions = { version = ">=4.4", optional = true }
rtoml = {version = "^0.9.0", optional = true}
pyyaml = {version = "^6.0", optional = true}

[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"
pre-commit = "*"
nonebug = "^0.3.1"
pytest-cov = "^4.0.0"
pytest-xdist = "^3.1.0"
pytest-mock = "^3.10.0"
pytest-asyncio = "^0.20.3"
nb-cli = "^1.0.0"
nonebot2 = { extras = ["httpx", "fastapi"], version = "^2.0.0-rc.1" }

[tool.poetry.extras]
cli = ["anyio", "click", "typing-extensions"]
toml = ["rtoml"]
yaml = ["pyyaml"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.pyright]
pythonVersion = "3.8"
pythonPlatform = "All"

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@overload",
]
omit = [
  "*/script/command.py",
]

[tool.poetry.plugins.nb_scripts]
datastore = "nonebot_plugin_datastore.script.cli:main"

[tool.nonebot]
plugins = ["nonebot_plugin_datastore"]
plugin_dirs = ["tests/registry"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
